skills/ai-data-engineering/examples/langchain-rag/requirements.txt
# LangChain RAG Example Dependencies

# Core LangChain
langchain-core==0.3.28
langchain==0.3.12

# LLM Providers
langchain-openai==0.2.14
langchain-anthropic==0.2.9
langchain-voyageai==0.2.0

# Vector Stores
langchain-qdrant==0.2.0
qdrant-client==1.12.1

# Utilities
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
